separate tests into e2e and unit
