+++ tests/e2e/actor-movement-validation.e2e.ts
