+++ tests/e2e/actor-pathfinding-validation.e2e.ts
