+++ tests/e2e/game-initialization.e2e.ts
