+++ tests/e2e/utils/canvasTestUtils.ts
