+++ tests/unit/discord-oauth.test.ts
@@ -1,5 +1,5 @@
 import { beforeEach, describe, expect, it, vi } from 'vitest';
-import { util } from '../src/websocket-utils.js';
+import { util } from '../../src/websocket-utils.js';
 import { MockLocalStorage } from './utils/testHelpers.js';
 
 // Create instance of mock localStorage
