+++ tests/unit/websocket-utils.test.ts
@@ -4,7 +4,7 @@
  */
 
 import { beforeEach, describe, expect, it, vi } from 'vitest';
-import { createJoinServerMessage, getStartupServer, util, type DiscordAuth, type GameState } from '../src/websocket-utils.js';
+import { createJoinServerMessage, getStartupServer, util, type DiscordAuth, type GameState } from '../../src/websocket-utils.js';
 import { MockLocalStorage } from './utils/testHelpers.js';
 
 // Create instance of mock localStorage
