+++ vitest.config.ts
@@ -3,8 +3,8 @@ import { defineConfig } from 'vitest/config';
 export default defineConfig({
   test: {
     /* Test file patterns */
-    include: ['tests/**/*.{test,spec}.{js,ts}'],
-    exclude: ['tests/**/*.e2e.{js,ts}'],
+    include: ['tests/unit/**/*.{test,spec}.{js,ts}'],
+    exclude: ['tests/e2e/**/*'],
     
     /* Global test setup */
     globals: true,
